Use long instead of int for View cursor position fields

+++ src/move.h
@@ -5,7 +5,7 @@
 #include "block-iter.h"
 #include "view.h"
 
-void move_to_preferred_x(int preferred_x);
+void move_to_preferred_x(long preferred_x);
 void move_cursor_left(void);
 void move_cursor_right(void);
 void move_bol(void);
